chore: refactor exports and imports
