+++ src/components/index.ts
@@ -1,0 +1,26 @@
+export { default as Aside } from './Aside/Aside';
+export { default as Button } from './Button/Button';
+export { default as ConfirmationPage } from './ConfirmationPage/ConfirmationPage';
+export { default as CustomCheckbox } from './CustomCheckbox/CustomCheckbox';
+export { default as CustomForm } from './CustomForm/CustomForm';
+export { default as CustomInput } from './CustomInput/CustomInput';
+export { default as CustomInputImage } from './CustomInputImage/CustomInputImage';
+export { default as CustomLabel } from './CustomLabel/CustomLabel';
+export { default as CustomRadioInput } from './CustomRadioInput/CustomRadioInput';
+export { default as CustomSelect } from './CustomSelect/CustomSelect';
+export { default as CustomTextarea } from './CustomTextarea/CustomTextarea';
+export { default as DropDownMenu } from './DropDownMenu/DropDownMenu';
+export { default as ErrorPage } from './ErrorPage/ErrorPage';
+export { default as ExpertCard } from './ExpertCard/ExpertCard';
+export { default as Experts } from './Experts/Experts';
+export { default as Filtration } from './Filtration/Filtration';
+export { default as Marquee } from './Marquee/Marquee';
+export { default as ModalWindow } from './ModalWindow/ModalWindow';
+export { default as ModalWindowForImage } from './ModalWindowForImage/ModalWindowForImage';
+export { default as Portal } from './Portal/Portal';
+export { default as ServiceCard } from './ServiceCard/ServiceCard';
+export { default as Services } from './Services/Services';
+export { default as Slider } from './Slider/Slider';
+export { default as SliderCard } from './SliderCard/SliderCard';
+export { default as Stats } from './Stats/Stats';
+export { default as Steps } from './Steps/Steps';
