+++ src/pages/FormPage/FormPage.tsx
@@ -8,9 +8,7 @@ import { FormValues, IFormikInnerRef } from '@/customTypes/formik.types';
 import { RootState } from '@/customTypes/store.types';
 import { IDataForForm } from '@/interfaces/IDataForForms';
 import { addItemToForm } from '@/store/actions/actions';
-import Button from '@components/Button/Button';
-import CustomForm from '@components/CustomForm/CustomForm';
-import Steps from '@components/Steps/Steps';
+import { Button, CustomForm, Steps } from '@/components';
 import { IFormikRef } from '@interfaces/IDataForFormik';
 import { Formik } from 'formik';
 import { motion } from 'framer-motion';
