+++ src/pages/TablePage/TablePage.tsx
@@ -1,5 +1,4 @@
-import Button from '@/components/Button/Button';
-import ModalWindow from '@/components/ModalWindow/ModalWindow';
+import { Button, ModalWindow, Filtration } from '@/components';
 import { EFormProps } from '@/customTypes/form.types';
 import { RootState } from '@/customTypes/store.types';
 import { IDataForForm } from '@/interfaces/IDataForForms';
@@ -9,9 +8,8 @@ import { ThemeContext } from '@/theme/theme';
 import FiltrationIcon from '@assets/icons/filtration.svg';
 import SortArrow from '@assets/icons/sort-arrow.svg';
 import cn from 'classnames';
-import { useContext, useEffect, useState } from 'react';
+import { ReactNode, useContext, useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import Filtration from '../../components/Filtration/Filtration';
 
 import style from './tablePage.module.scss';
 
@@ -201,7 +199,7 @@ function TablePage() {
 									</td>
 									<td className={style.table__descr}>{item.comments}</td>
 									<td className={style.table__descr}>{item.profilePicture}</td>
-									<td className={style.table__descr}>{item.date}</td>
+									<td className={style.table__descr}>{item.date.toLocaleString()}</td>
 								</tr>
 							))
 						) : (
